Move checkCanOverwrite to its own file
