+++ src/utils/checkCanOverwrite.ts
@@ -1,0 +1,49 @@
+import { DialogResponses } from "vscode-azureextensionui";
+import { ext } from "../extensionVariables";
+import { localize } from "./localize";
+import { OverwriteChoice } from "./uploadUtils";
+
+// Pass `overwriteChoice` as an object to make use of pass by reference.
+export async function checkCanOverwrite(
+    destPath: string,
+    overwriteChoice: { choice: OverwriteChoice | undefined },
+    destPathExists: () => Promise<boolean>
+): Promise<boolean> {
+    if (overwriteChoice.choice === OverwriteChoice.yesToAll) {
+        // Always overwrite
+        return true;
+    }
+
+    if (await destPathExists()) {
+        if (overwriteChoice.choice === OverwriteChoice.noToAll) {
+            // Resources that already exist shouldn't be overwritten
+            return false;
+        } else {
+            overwriteChoice.choice = await showDuplicateResourceWarning(destPath);
+            switch (overwriteChoice.choice) {
+                case OverwriteChoice.no:
+                case OverwriteChoice.noToAll:
+                    return false;
+
+                case OverwriteChoice.yes:
+                case OverwriteChoice.yesToAll:
+                default:
+                    return true;
+            }
+        }
+    } else {
+        // This resource doesn't exist yet, so overwriting is OK
+        return true;
+    }
+}
+
+async function showDuplicateResourceWarning(resourceName: string): Promise<OverwriteChoice> {
+    const message: string = localize('resourceExists', 'A resource named "{0}" already exists. Do you want to overwrite it?', resourceName);
+    const items = [
+        { title: localize('yesToAll', 'Yes to all'), data: OverwriteChoice.yesToAll },
+        { title: DialogResponses.yes.title, data: OverwriteChoice.yes },
+        { title: localize('noToAll', 'No to all'), data: OverwriteChoice.noToAll },
+        { title: DialogResponses.no.title, data: OverwriteChoice.no }
+    ];
+    return (await ext.ui.showWarningMessage(message, { modal: true }, ...items)).data;
+}
